Archive unused CacheOpendapStation and unused parts of SSR.
